+++ Pinionszek_API/Services/DatabaseServices/BudgetService/BudgetApiService.cs
@@ -42,14 +42,6 @@ namespace Pinionszek_API.Services.DatabaseServices.BudgetService
             return budget;
         }
 
-        public async Task<SharedPayment?> GetSharedPaymentDataAsync(int idPayment)
-        {
-            return await _dbContext.SharedPayments
-                .Where(sp => sp.IdPayment == idPayment)
-                .FirstOrDefaultAsync();
-
-        }
-
         public async Task<(string?, int?)> GetFriendReceiveNameAndTagAsync(int idSharedPayment)
         {
             var friendReceiverQuery = await (from sp in _dbContext.SharedPayments
@@ -82,57 +74,6 @@ namespace Pinionszek_API.Services.DatabaseServices.BudgetService
                                            ).FirstOrDefaultAsync();
 
             return (friendSenderQuery?.Login, friendSenderQuery?.UserTag);
-        }
-
-        public async Task<IEnumerable<Payment>> GetAssignedPaymentsAsync(int friendTag)
-        {
-            return await (from sh in _dbContext.SharedPayments
-                          join f in _dbContext.Friends
-                          on sh.IdFriend equals f.IdFriend
-
-                          where f.FriendTag == friendTag
-
-                          join p in _dbContext.Payments
-                          on sh.IdPayment equals p.IdPayment
-
-                          join ps in _dbContext.PaymentStatuses
-                          on p.IdPaymentStatus equals ps.IdPaymentStatus
-
-                          join dc in _dbContext.DetailedCategories
-                          on p.IdDetailedCategory equals dc.IdDetailedCategory
-
-                          join gc in _dbContext.GeneralCategories
-                          on dc.IdGeneralCategory equals gc.IdGeneralCategory
-
-                          select new Payment
-                          {
-                              IdPayment = p.IdPayment,
-                              Name = p.Name,
-                              Charge = p.Charge,
-                              Refund = p.Refund,
-                              Message = p.Message,
-                              PaymentDate = p.PaymentDate,
-                              PaidOn = p.PaidOn,
-                              PaymentAddedOn = p.PaymentAddedOn,
-                              IdPaymentStatus = p.IdPaymentStatus,
-                              PaymentStatus = p.PaymentStatus,
-                              IdDetailedCategory = p.IdDetailedCategory,
-                              SharedPayment = (from sp in _dbContext.SharedPayments
-                                               where sp.IdPayment == p.IdPayment
-                                               select sp).FirstOrDefault(),
-                              DetailedCategory = new DetailedCategory
-                              {
-                                  IdDetailedCategory = dc.IdDetailedCategory,
-                                  IdGeneralCategory = dc.IdGeneralCategory,
-                                  Name = dc.Name,
-                                  GeneralCategory = new GeneralCategory
-                                  {
-                                      IdGeneralCategory = gc.IdGeneralCategory,
-                                      Name = gc.Name,
-                                      IsDefault = gc.IsDefault
-                                  }
-                              }
-                          }).ToListAsync();
         }
 
         public async Task<UserSettings?> GetUserSettingsAsync(int idUser)
