+++ Pinionszek_API/Services/DatabaseServices/BudgetService/IBudgetApiService.cs
@@ -6,10 +6,8 @@ namespace Pinionszek_API.Services.DatabaseServices.BudgetService
     public interface IBudgetApiService
     {
         public Task<Budget?> GetBudgetDataAsync(int idUser, DateTime budgetYear);
-        public Task<SharedPayment?> GetSharedPaymentDataAsync(int idPayment);
         public Task<(string?, int?)> GetFriendReceiveNameAndTagAsync(int idSharedPayment);
         public Task<(string?, int?)> GetFriendSenderNameAndTagAsync(int idSharedPayment);
-        public Task<IEnumerable<Payment>> GetAssignedPaymentsAsync(int friendTag);
         public Task<UserSettings?> GetUserSettingsAsync(int idUser);
         public Task<IEnumerable<Budget>> GetBudgetsAsync(int idUser);
         public Task<Payment?> GetPaymentAsync(int idPayment, int idUser);
