+++ Pinionszek_API/Services/DatabaseServices/PaymentService/IPaymentApiService.cs
@@ -5,5 +5,7 @@ namespace Pinionszek_API.Services.DatabaseServices.PaymentService
     public interface IPaymentApiService
     {
         public Task<IEnumerable<Payment>> GetPaymentsAsync(int idBudget);
+        public Task<IEnumerable<Payment>> GetAssignedPaymentsAsync(int friendTag);
+        public Task<SharedPayment?> GetSharedPaymentDataAsync(int idPayment);
     }
 }
